from Long to Integer, deop StatoEntity

+++ progetto-fermi-bean/src/main/java/it/previnet/progettofermi/bean/intf/WithToken.java
@@ -2,5 +2,5 @@ package it.previnet.progettofermi.bean.intf;
 
 public interface WithToken {
 
-    Long getToken();
+    Integer getToken();
 }
